renamed loads of(2) files for easier autoloading
