+++ index.php
@@ -1,4 +1,4 @@
 <?php
-    require_once "app/database/dbDriverMySQL.php";
+    require_once "app/drivers/MySQLDriver.php";
 
     print_r(dbDriverMySQL::query("SELECT * FROM `test`", []));
